fuzzy/requirements.txt
blinker==1.3
decorator==3.4.2
Flask==0.10.1
Flask-Login==0.2.11
Flask-Mail==0.9.1
Flask-OpenID==1.2.4
Flask-SQLAlchemy==2.0
Flask-WhooshAlchemy==0.56
Flask-WTF==0.11
itsdangerous==0.24
Jinja2==2.7.3
MarkupSafe==0.23
pbr==0.11.0
python-openid==2.2.5
six==1.9.0
SQLAlchemy==0.9.9
sqlalchemy-migrate==0.9.6
sqlparse==0.1.15
Tempita==0.5.2
Werkzeug==0.10.4
Whoosh==2.7.0
WTForms==2.0.2
